requirements.txt
Django==4.1.3
google_api_python_client==2.74.0
google_auth_oauthlib==0.8.0
protobuf==4.21.12
pymongo==4.3.3
django-cors-headers==3.13.0
gunicorn==20.1.0
pandas==1.5.3
numpy==1.24.1